Major overhault of userProject
