+++ web/src/style/style.css
@@ -208,8 +208,9 @@ a:hover {
 /* Profile Details */
 
 .profile-header {
-  margin-top: 30px;
-  margin-bottom: 10px;
+  margin-bottom: 30px;
+  margin-top: 20px;
+
 }
 
 .profile-info {
@@ -220,16 +221,17 @@ a:hover {
   font-weight: bold;
 }
 
+.bio-info {
+  margin-top: 20px;
+}
+
 .profile-icon {
-  width: 200px;
-  height: 200px;
+  width: 180px;
+  height: 180px;
   border-style: solid;
   border-width: 1px;
   border-color: grey;
   display: block;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 30px;
 }
 
 .change-password {
